Part object is nearing completion
I THINK that I was able to write a converter in Ruby that will convert a
folder of .mid files to a .tar.gz of text files. We will see when I move
on some more again. Need to implement OutputPart next, possibly edit
Part a little so that it does not contain an ArrayList<OutputPart> but
rather an ArrayList<Long[]> for the times of output queues, and then
read M and P.