Started output.
Pushing phrase matching to the back-burner for now so that I can meet
the deadline. Currently starting output to MIDI ports. Sending notes at
time 0 when Receiver is connected. Then, waiting for time updates. When
all parts are at or past the time for the next event, output that event
and wait for the next one.