Started working on timing output.
PlayThread also works as clock now, it is the only one that sends
messages to output threads. It keeps track of parts and creates a time
multiplier to determine where exactly we are in a song relative to where
the system clock is.

Need to work on continuing implementation in output so that it handles
these messages.